Rename lib/output/ to lib/output-file/
