+++ lib/index.js
@@ -12,7 +12,7 @@ module.exports = {
   ...require('./mkdirs'),
   ...require('./move-sync'),
   ...require('./move'),
-  ...require('./output'),
+  ...require('./output-file'),
   ...require('./path-exists'),
   ...require('./remove')
 }
