+++ lib/json/output-json-sync.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { stringify } = require('jsonfile/utils')
-const { outputFileSync } = require('../output')
+const { outputFileSync } = require('../output-file')
 
 function outputJsonSync (file, data, options) {
   const str = stringify(data, options)
