+++ lib/json/output-json.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const { stringify } = require('jsonfile/utils')
-const { outputFile } = require('../output')
+const { outputFile } = require('../output-file')
 
 async function outputJson (file, data, options = {}) {
   const str = stringify(data, options)
