+++ lib/output-file/index.js
